+++ web/src/Pages/ProgrammeManagement/ProjectDetailsView.tsx
@@ -2,7 +2,7 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import SLCFProjectDetailsViewComponent from '../../Components/ProjectDetailsView/SLCFProjectDetailsViewComponent';
 
-const SLCFProjectDetailsView = () => {
+const ProjectDetailsView = () => {
   const navigate = useNavigate();
   const { i18n } = useTranslation(['common', 'projectDetailsView']);
 
@@ -18,4 +18,4 @@ const SLCFProjectDetailsView = () => {
   );
 };
 
-export default SLCFProjectDetailsView;
+export default ProjectDetailsView;
