+++ web/src/Pages/ProgrammeManagement/programmeManagement.tsx
@@ -2,47 +2,48 @@ import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { useAbilityContext } from '../../Casl/Can';
 import { ProgrammeManagementColumns } from '../../Definitions/Enums/programme.management.columns.enum';
-import { ProgrammeManagementComponent } from '../../Components/Programme/ProgrammeManagement/programmeManagementComponent';
+import { ProgrammeManagementSlColumns } from '../../Definitions/Enums/programme.management.sl.columns.enum';
+import { SLCFProgrammeManagementComponent } from '../../Components/SLCFProgrammeManagement/SLCFProgrammeManagementComponent';
 
 const ProgrammeManagement = () => {
   const navigate = useNavigate();
-  const { t } = useTranslation(['common', 'programme']);
+  const { t } = useTranslation(['common', 'projectList']);
 
   const visibleColumns = [
-    ProgrammeManagementColumns.title,
-    ProgrammeManagementColumns.company,
-    ProgrammeManagementColumns.sector,
-    ProgrammeManagementColumns.currentStage,
-    ProgrammeManagementColumns.creditBalance,
-    ProgrammeManagementColumns.creditTransferred,
-    ProgrammeManagementColumns.certifierId,
-    ProgrammeManagementColumns.serialNo,
-    ProgrammeManagementColumns.emissionReductionExpected,
-    ProgrammeManagementColumns.emissionReductionAchievedandCreditIssued,
-    ProgrammeManagementColumns.action,
+    ProgrammeManagementSlColumns.title,
+    ProgrammeManagementSlColumns.company,
+    ProgrammeManagementSlColumns.projectCategory,
+    ProgrammeManagementSlColumns.projectProposalStage,
+    ProgrammeManagementSlColumns.projectStatus,
+    ProgrammeManagementSlColumns.creditBalance,
+    ProgrammeManagementSlColumns.purposeOfCreditDevelopment,
+    ProgrammeManagementSlColumns.creditRetired,
+    ProgrammeManagementSlColumns.certifierId,
+    ProgrammeManagementSlColumns.serialNo,
+    ProgrammeManagementSlColumns.action,
   ];
 
   const onNavigateToProgrammeView = (record: any) => {
-    navigate(`/programmeManagement/view/${record.programmeId}`, { state: { record } });
+    navigate(`/programmeManagementSLCF/view/${record.id}`, { state: { record } });
   };
 
   const onClickAddProgramme = () => {
-    navigate('/programmeManagement/addProgramme');
+    navigate('/programmeManagementSLCF/addProgramme');
   };
 
   const onClickAddInvestment = () => {
-    navigate('/programmeManagement/addInvestment');
+    navigate('/programmeManagementSLCF/addInvestment');
   };
 
   return (
-    <ProgrammeManagementComponent
+    <SLCFProgrammeManagementComponent
       t={t}
       visibleColumns={visibleColumns}
       onNavigateToProgrammeView={onNavigateToProgrammeView}
       onClickAddProgramme={onClickAddProgramme}
       enableAddProgramme
       useAbilityContext={useAbilityContext}
-    ></ProgrammeManagementComponent>
+    ></SLCFProgrammeManagementComponent>
   );
 };
